test: move helper functions in respective test suites
